Refactor: move history_db.py to chatbot module
- Move `plugin/framework/history_db.py` to `plugin/modules/chatbot/history_db.py` to better reflect its domain ownership.
- Update import paths in `smolagents/default_tools.py`, `legacy_ui.py`, `chatbot/panel.py`, and `test_history_db.py`.
- Update references in `docs/audio-architecture.md` and `AGENTS.md`.
