+++ AGENTS.md
@@ -132,7 +132,7 @@ The sidebar and menu Chat work for **Writer and Calc** (same deck/UI; ContextLis
 - **Chat** uses `get_chat_system_prompt_for_document(model, additional_instructions)` in `plugin/framework/constants.py` so the correct prompt is chosen by document type: **Writer** → `DEFAULT_CHAT_SYSTEM_PROMPT` + additional_instructions (get_markdown/apply_markdown, presume document editing, translate/proofread, no preamble); **Calc** → `DEFAULT_CALC_CHAT_SYSTEM_PROMPT` + additional_instructions (semicolon formula syntax, 4-step workflow: understand → get state → use tools → short confirmation; tools grouped READ / WRITE & FORMAT / SHEET MANAGEMENT / CHART / ERRORS). Used by both sidebar and menu Chat.
 - **Reasoning tokens**: `plugin/main.py` sends `reasoning: { effort: 'minimal' }` on all chat requests.
 - **Thinking display**: Reasoning tokens are shown in the response area as `[Thinking] ... /thinking`. When thinking ends we append a newline after ` /thinking` so the following response text starts on a new line.
-- **Persistent Chat History**: Logic in `plugin/framework/history_db.py`. Uses native `sqlite3` for chat history persistence when available; **SQLite is optional**: if `sqlite3` is not available (e.g. some Windows/bundled Python builds), history falls back to JSON files in `writeragent_history.db.d/`. Global guard: `plugin/framework/sqlite_available.py` (`HAS_SQLITE`).
+- **Persistent Chat History**: Logic in `plugin/modules/chatbot/history_db.py`. Uses native `sqlite3` for chat history persistence when available; **SQLite is optional**: if `sqlite3` is not available (e.g. some Windows/bundled Python builds), history falls back to JSON files in `writeragent_history.db.d/`. Global guard: `plugin/framework/sqlite_available.py` (`HAS_SQLITE`).
     - **Schema**: Simple `message_store` table compatible with LangChain's SQL history JSON format.
     - **Database Path**: Stored in LibreOffice user config directory (`writeragent_history.db`).
     - **Robust Session IDs**: Instead of document URLs (which break on rename), we store a `WriterAgentSessionID` in the document's **`UserDefinedProperties`** via `getDocumentProperties()`. If missing, we generate one (hash or UUID) and persist it to the file. This ensures history remains linked even if the file is moved or renamed.
