+++ plugin/framework/legacy_ui.py
@@ -20,7 +20,7 @@ from plugin.framework.uno_context import get_desktop, get_active_document, get_e
 from plugin.framework.dialogs import TabListener, is_checkbox_control, get_checkbox_state, set_checkbox_state, get_optional, set_control_enabled, set_control_text, get_control_text
 from plugin.framework.config import get_config, get_current_endpoint, get_text_model, populate_combobox_with_lru, set_config, update_lru_history
 from plugin.framework.logging import init_logging, agent_log
-from plugin.framework.history_db import HAS_SQLITE
+from plugin.modules.chatbot.history_db import HAS_SQLITE
 import uno
 
 log = logging.getLogger(__name__)
